Fix `GET method cannot be used with a body` error
During our testing we thought we were okay to send a body in a GET request providing the body was an empty object `{}`. However when updating `InfoService` we realised that this causes an error as it still tries to add a body to the request. We therefore remove our default of `{}` so that for a GET request, the body is `undefined` which is acceptable when making a request.