Send request payload as text instead of json
Previously we were sending our request payload as json rather than as text. Doing this switches got into "json mode", where the response body is returned as json. However as we don't pass a payload for GET requests, this meant that our POST requests were coming back with a json body but our GET requests had a text body. Even though we then passed it through JSON.parse() to get json, this meant we were effectively passing an object through this which isn't ideal. We therefore change how we send the payload so that we always send it as text, always get text back, and therefore always pass text through JSON.parse().