Leave response body as-is when creating bill run
When we create a bill run, the CM's response body contains a `billRun` object. Previously, we took the contents of this object and overwrote the entire response with it. This meant that we then lost things like headers, status code, etc. and were not consistent with having the response `body` be the place to look for the returned data.

To ensure consistency we therefore no longer overwrite the response.