Spike - Create a complete debit only SROC bill run
https://eaflood.atlassian.net/browse/WATER-3906

We have created the individual 'tools' needed to generate an SROC supplementary bill-run. By this, we mean the `billing_batch`, `billing_invoice`, `billing_invoice_licence` and `billing_transaction` records. Also, (though we're still not sure where it's used!) the event record.

We have also worked to identify the charge versions to consider in an SROC supplementary bill-run, and how to communicate with the charging module API. Now it's time to pull it all together and create a complete bill-run, one that can be confirmed and sent in the WRLS service UI.

The work to credit a previous bill run is still to be done. But we have everything we need to get us to a place where we folks can start testing and verifying debit-only bill-runs.

So, we're starting by spiking a solution. We don't intend to merge this PR. But we need some space to work out how to hang together a solution from which we'll then extract our proper changes.