Just status code & body - but we broke it!
By adding the **Got** [responsetype](https://github.com/sindresorhus/got/blob/main/documentation/2-options.md#responsetype) option and setting it to 'json' it automatically parses the response body for us.

We also tweaked the `_parseResult()` method to just return the things from the response we need, or so we thought!

We were using the `/health/info` to quickly test the changes and when we made this we broke it! That's because `app/services/health/info.service.js` expects to be able to access the headers. So, we're going to need to tweak our solution to include that as well.