Handle Charging Module Headers
The charging module sends back 2 custom headers. These are used to denote the specific version of the code running, and which Docker image is being used.

These are all we care about in the headers. So, rather than just plonk all the headers in our response we use the method to do the work of extracting them for easier access by our services.