Replace body with json in options
We've had issues with this before. But we think we have cracked it now.

If you set the **Got** option [json:](https://github.com/sindresorhus/got/blob/main/documentation/2-options.md#json) then it _must_ be a JavaScript JSON object and **Got** will handle the stringify.

If you set [body:](https://github.com/sindresorhus/got/blob/main/documentation/2-options.md#body) then it _must_ be a string and you cannot use the `json:` option.

Where we've had it wrong before is because we were either setting `body:` to a JSON object, or passing a stringified JSON object to `json:`.