Fix response handling for Charging Module requests (#138)
https://eaflood.atlassian.net/browse/WATER-3906

We've spotted that we're not being consistent with our response handling of requests to the [sroc-charging-module-api](https://github.com/DEFRA/sroc-charging-module-api).

We'd done some refactoring in [Refactor to use ChargingModuleRequestLib](https://github.com/DEFRA/water-abstraction-system/pull/130) to the `_parseResult()` function of `app/lib/charging-module-request.lib.js`.

The result was a calling service would receive something like

```javascript
{
  succeeded: true,
  response: {
    body: {
      billRun: {
        id: ...,
        billRunNumber: ...
      }
    },
    statusCode: ...
  }
}
```

This started blowing up when we started work on the [Spike - Create a complete debit only SROC bill run](https://github.com/DEFRA/water-abstraction-system/pull/133). This was because `app/services/supplementary-billing/initiate-billing-batch.service.js` was trying to access the result as below

```javascript
const options = {
  scheme,
  batchType: type,
  externalId: chargingModuleBillRun.response.id
}
```

To resolve this we came to the conclusion that, exceptions aside, what we want from [Got](https://github.com/sindresorhus/got) is the response body as is, plus the status code. That's it. This applies to `2xx/3xx` and `4xx/5xx` responses.

This change updates `app/lib/charging-module-request.lib.js` to ensure this.

We could have done it in `app/lib/request.lib.js` but we will be using **Got** options that may not be applicable in all cases. So, by doing it at the `charging-module-request.lib.js` level we retain flexibility in how we make requests should we need it.

---

The bulk of this work involved tweaking `_parseResult()` to just return the things from the response we need, or so we thought! `app/services/health/info.service.js` expects to be able to access the headers returned from the Charging Module. They are 2 custom headers used to denote the specific version of the code running, and which Docker image is being used. So, we expanded our solution to include those as well.

We also took the opportunity to make use of the **Got's** inbuilt functionality to do things like parse data to and from the external service, and formulate the full URL to use. This reduces the work we're doing, and the complexity of our solution.

As part of this we cracked our confusion with `body` vs `json`.

If you set the **Got** option [json:](https://github.com/sindresorhus/got/blob/main/documentation/2-options.md#json) then it _must_ be a JavaScript JSON object you set in the options and **Got** will handle the stringify.

If you set [body:](https://github.com/sindresorhus/got/blob/main/documentation/2-options.md#body) then it _must_ be a string and you cannot use the `json:` option.

Where we've had it wrong before is because we were either setting `body:` to a JSON object, or passing a stringified JSON object to `json:`.

Other than that it was updating or fixing services that use `app/lib/charging-module-request.lib.js`.