Add Charging Module create transaction service in water-abstraction-system
https://eaflood.atlassian.net/browse/WATER-3911

The transactions we generate in the WRLS service as part of the SROC supplementary bill run need to be added to the Charging Module’s bill run. We need to add a service that knows how to communicate with the charging module to send the request and handle the response.
- add the new `CreateTransactionService`
- add unit tests for the service