Fix missing bill run number in batch (#137)
https://eaflood.atlassian.net/browse/WATER-3906

We spotted during testing that the success page after confirming an SROC supplementary bill run displayed an `undefined` (This is in the [water-abstraction-ui](https://github.com/DEFRA/water-abstraction-ui)). What it should have been showing was the bill run number.

This comes from the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api) in the response after creating the bill run. We found the legacy code saves this but we've overlooked it in our version.

This change fixes this oversight.