Add Charging Module create transaction service (#140)
https://eaflood.atlassian.net/browse/WATER-3911

The transactions we generate in the WRLS service as part of the SROC supplementary bill run need to be added to the Charging Module’s bill run. We already created the bill run using `app/services/charging-module/create-bill-run.service.js`. So, following that pattern, we add `app/services/charging-module/create-transaction.service.js` in this change.

---

**Housekeeping**

It is a personal preference that has become a convention in our tests. Executing the thing under test should be done within the test rather than a `before()` block.

It's our opinion when this isn't done you lose sight of what a test is testing as you add more and more unit tests. It does mean you get some duplication. But we think it makes for more readable code.

Because we based on unit tests on those in `test/services/charging-module/create-bill-run.service.test.js` we spotted this deviation and corrected it as part of this change.