Housekeeping
It is a personal preference that has become a convention in our tests. Executing the thing under test is done within the test rather than a `before()` block.

IMHO when this isn't done it is easy to lose sight of what a test is testing as you add more and more unit tests. It does mean you get some duplication. But we think it makes for more readable code.