Add PATCH support to our request lib modules (#142)
https://eaflood.atlassian.net/browse/WATER-3906
https://eaflood.atlassian.net/browse/WATER-3919

In order to call the [/generate](https://defra.github.io/sroc-charging-module-api-docs/#/bill-run/GenerateBillRun) bill run endpoint in the Charging Module API we need to be able to make `PATCH` HTTP requests.

This change adds the functionality to `app/lib/request.lib.js` and `app/lib/charging-module-request.lib.js`.

---

**Housekeeping**

It is a personal preference that has become a convention in our tests. Executing the thing under test should be done within the test rather than a `before()` block.

It's our opinion when this isn't done you lose sight of what a test is testing as you add more and more unit tests. It does mean you get some duplication. But we think it makes for more readable code.

As we were adding new tests to `test/lib/charging-module-request.lib.test.js` we spotted this deviation and corrected it as part of this change.