Add new legacy auth token env var
The legacy services use a shared [JSON Web Token (JWT)](https://jwt.io/) which is saved as an env var in their own config files. Requests need to have this set in their `authorisation` header or they'll be rejected.

TBH, it is a pointless authentication mechanism! All the legacy services share the same token, and read it from config files stored on the server instances. From a security point of view, it's the same as keeping your password on a post-it note stuck to your monitor.

But, Rome wasn't built in a day. So, for now, our new system has to play along.