Add LegacyRequestLib (#143)
https://eaflood.atlassian.net/browse/WATER-3906
https://eaflood.atlassian.net/browse/WATER-3921

For the SROC supplementary bill run process we intend to take things as far as having the batch, invoice, invoice licence and transaction records created in the DB and the Charging Module API, and the bill run generated.

Control will then be handed to the legacy service to complete the process. `src/modules/billing/jobs/create-charge.js` in the [water-abstraction-service](https://github.com/DEFRA/water-abstraction-service) has been our guide for this. After generating the bill run in the charging module it queues up the `src/modules/billing/jobs/refresh-totals.js` job.

We've [added a new endpoint](https://github.com/DEFRA/water-abstraction-service/pull/2036) to the **water-abstraction-service** that allows us to do the same. The first step is to add a new `LegacyRequestLib` following the same pattern as `app/lib/charging-module-request.lib.js`.

---

As part of this change, we have to add a new env var to support talking to the legacy services. We also update our existing `app/services/health/info.service.js` to make use of `LegacyRequestLib`.