Add Base Presenter module
https://eaflood.atlassian.net/browse/WATER-3906
https://eaflood.atlassian.net/browse/WATER-3925

In the [sroc-charging-module-api](https://github.com/DEFRA/sroc-charging-module-api) project all our 'presenters' inherited from a `base.presenter.js`. It contained handy util methods, for example, `formatDate()` which formats a date as a `DD-MMM-YYYY` string.

Turns out, we need exactly that when preparing the transaction data to be sent to the charging module API! So, we add a `base.presenter.js` module for the first time. We'll bring over what we have in the **sroc-charging-module-api** as it has what we need including unit tests.

One thing to note is the **sroc-charging-module-api** was built mainly from [classes](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Classes). We wanted to align more with JavaScript conventions for this project so have stuck to modules. That means our presenters won't inherit the base. Instead those that need it can just `require()` it and its functions. We've stuck with the name 'base' though as its our convention for a base object that is intended to be used by related objects.