Add Charging Module Transaction Request Presenter
https://eaflood.atlassian.net/browse/WATER-3906
https://eaflood.atlassian.net/browse/WATER-3926

It is not a one-to-one mapping of transaction properties between our service and the [Charging Module API](https://github.com/DEFRA/sroc-charging-module-api). When we have data that we need to 'present' in a certain way we use the concept of presenters.

As part of the [spike work](https://github.com/DEFRA/water-abstraction-system/pull/133) we created one that 'presents' a WRLS transaction in the format needed to send as a request to the Charging Module API.

This adds it to the project properly along with unit tests.