Return PostgreSQL decimals as numbers not strings (#146)
https://eaflood.atlassian.net/browse/WATER-3906
https://eaflood.atlassian.net/browse/WATER-3927

With our [Add billing_transaction model to water-abstraction-system](https://github.com/DEFRA/water-abstraction-system/pull/135) change we had our first decimal field.

When we created the migration, we didn't realise that the [pg package](https://www.npmjs.com/package/pg) returns decimals as strings.

We've encountered this issue with BigInt on the [charging-module-api](https://github.com/DEFRA/sroc-charging-module-api) which is why `db/db.js` has logic to parse the string **pg** returns back into an integer.

The root cause is the same. Out of concern about the loss of precision the DB values are returned as strings. We know we're not dealing with values where this would be an issue so we can have **pg** implicitly parse the decimal strings into floats.

[This post](https://stackoverflow.com/a/39176670/6117745) focuses on the BigInt issue. But the cause and solution are the same for decimals.

So, this change ensures when we use [Knex](https://github.com/knex/knex) or [Objection.js](https://github.com/vincit/objection.js) we see floats (numbers) instead of strings.

---

Whilst making this change we also updated the existing fix for BigInts to use the same solution. Using named types rather than a 'magic number' from the DB was preferable in our opinion