Add billing period in InitiateBillingBatch result
https://eaflood.atlassian.net/browse/WATER-3906
https://eaflood.atlassian.net/browse/WATER-3929

To create the billing invoice, invoice licence, and transaction records during an SROC supplementary bill run we need to know the billing period we're generating the bill run for. This is determined in the `InitiateBillingBatchService` but not returned in its result. This change adds billing period to the result so we can it pass onto a new `ProcessBillingBatchService` (coming soon!)