Add new ProcessBillingBatchService as placeholder
This is what we have been working on in a spike and intend to be the service which manages the process of populating the bill run, the charging module's bill run, and handing the process back to the legacy code.

We're adding this first so we can then call it from our InitiateBillingBatchService.