Start processing the billing batch from Initiate (#147)
https://eaflood.atlassian.net/browse/WATER-3906
https://eaflood.atlassian.net/browse/WATER-3929

Our `InitiateBillingBatchService` handles creating the initial WRLS and Charging Module bill runs. It then returns information about both to the [water-abstraction-service](https://github.com/DEFRA/water-abstraction-service) which makes the create bill run request.

Using the same design as the [charging-module-api](https://github.com/DEFRA/sroc-charging-module-api), our intent is to perform a small action like this, and then return the result in less than a second. As the processing of the billing batch will take more than a second, this must happen in the background and _not_ be something a user is left waiting for.

But we still need to kick it off and as the `InitiateBillingBatchService` has all the data we need to do that this change updates the service to call a new `ProcessBillingBatchService`.

> Note: `ProcessBillingBatchService` will just be a placeholder for now; it'll do something to prove the design works and we'll update it later.