Add Determine Minimum Charge Service
https://eaflood.atlassian.net/browse/WATER-3906
https://eaflood.atlassian.net/browse/WATER-3928

There is a concept of minimum charge. This is applied where the charge for certain customers falls below £25. The Charging Module will add an extra transaction line to the invoice to make the charge up to £25.

From looking at the legacy code there are some tests that get applied to determine if a charge version is subject to minimum charge. So, this change adds a service to perform these tests and return a `true|false`. We'll then use this information in our `ProcessBillingBatchService`.