Move determineChargePeriod() to a service
https://eaflood.atlassian.net/browse/WATER-3906
https://eaflood.atlassian.net/browse/WATER-3930

The charge period is the period of time that the charge version and its start and end date overlaps with the financial period being billed (01-Apr to 31-Mar). We need this period to then calculate how many billable days there are for the charge version.

Currently, the `FormatSrocTransactionLineService` has this function. But we also need to know the charge period to [determine whether minimum charge applies](https://github.com/DEFRA/water-abstraction-system/pull/149) to the charge version.

So, this change moves the function out to a service we can then reuse in both places.