Refactor CreateTransactionsService to persist (#151)
https://eaflood.atlassian.net/browse/WATER-3906
https://eaflood.atlassian.net/browse/WATER-3931

As part of the spike work, we are doing to get the complete SROC supplementary billing process up and running we realised that the existing `CreateTransactionsService` and `FormatSrocTransactionLineService` are doing very similar things. We also realised that determining how many transaction lines to create needs to be done within the main `ProcessBillingBatchService`.

But we do still need something that will persist a transaction to the DB. So, this change refactors the service to be that thing!