Rename service
Dropping the SROC because SROC is always assumed. Should we take on PRESROC functionality _then_ we'll label it as such.

Replacing format with generate as we are 'generating' the transaction data from other data elements. It is also becoming our convention where a service is responsible for extracting data from other things to create a new object, but not persist it.