Amend abstraction period logic
Our function for calculating abstraction periods didn't work when the reference period started and ended the same year. We have amended the logic so that instead of checking the start and end day/month and creating dates accordingly, we simply create a start and end date using the reference period's start year, and fix the end date's year if needed. Doing this means we also need to create an additional abstraction period to consider but we will look at refactoring to only do this if necessary