Create chargePeriod outside the service
The charge period is based on the charge version and the billing period. So, when called within the transaction service we are just getting the same result.

To avoid executing unnecessary code, we create the charge period outside the service and pass it in. This also means we can drop charge version as an arg to the service.