Alan's suggested edit
The size of the `go()` function was concerning me at this point. I was also weighing up the trade-offs between keeping another list of objects in memory that could potentially be as large as the original `generatedInvoices` vs the overhead of calling `filter()` on our original list.

My suggestion is to go with the latter, because it also helps shift some of the logic out of the `go()` function so we can break down what is happening and IMHO make things a little clearer.

But this is a suggestion so I am happy to be argued out of it. 🤷