Handle errors in ProcessBillingBatchService
https://eaflood.atlassian.net/browse/WATER-3924

Should an error occur during the supplementary bill run process it brings down the service. We're not handling the exception and because we moved the process to the background in our bill runs controller, there is nothing to deal with it when thrown.

So, we need to protect the service from _any_ errors during the process, to prevent the service going down. But we also need to mimic the legacy service and use an 'error code' to provide extra information about what the error was. We've already used this concept with [Create error bill run when CM fails](https://github.com/DEFRA/water-abstraction-system/pull/104) where we apply a code to the `billing_batch` record which the UI then uses to indicate what happened to the user.

This change deals with both; ensuring we handle any exceptions raised and where possible, set a relevant error code when updating the `billing_batch` status to 'errored'.