Log time taken to process a bill run
https://eaflood.atlassian.net/browse/WATER-3906

When issues of performance came up in the [charging-module-api](https://github.com/DEFRA/sroc-charging-module-api) we found it immensely useful to have evidence to hand of how long the most contentious process was actually taking.

As we are not in control of the full process with our new SROC supplementary billing process, it would be useful to know exactly how long are part of the journey is taking. We're also about to expand the process to include reversing the previous supplementary bill run. So, it would be useful to get a working idea of how that impacts our timings.

To be clear, this is a very basic metric; the difference between the times at the start and the end of the process. There is no call at this time for factoring the size of a bill run, load on the servers at the time etc. It's all we needed in the Charging Module API, which also means it's some code we can readily steal!