Fix cached data in ProcessBillingBatchService
https://eaflood.atlassian.net/browse/WATER-3906

As part of implementing [Handle errors in ProcessBillingBatchService](https://github.com/DEFRA/water-abstraction-system/pull/161) we refactored the `ProcessBillingBatchService` to make the code cleaner and easier to read.

One of the changes was moving the `generatedInvoices` and `generatedInvoiceLicences` to the module scope rather than just in the method scope. It saved us having to pass them around the place which removed a bunch of 'cruft'. What we didn't take account of is they would be cached between calls to the service.

Now it's at module scope the data from previous runs is being cached and still exists when a new one starts. So, we're getting to the finalise stage and we're trying to persist old data from a previous run!

This change resolves the issue.