Implement fix
We move the information into the scope of the `go()` function removing the issue of the data being cached at the global scope.

We then take a half-way house approach; we pass the information into our methods but we accept having to mutate the the data rather than pass back a modified copy.