Create Reverse Billing Batch Licences service
https://eaflood.atlassian.net/browse/WATER-3936

When we generate a supplementary bill run we are required to reverse the previous one. Essentially, when a licence is flagged for supplementary billing, for whatever reason, we reverse any previous bill runs and generate a new one.

With the exception of Annual billing, because we know any previous supplementary bill run will have reversed the one that came before, any new runs also just need to focus on the ones previous to them.

So, for each charge version `FetchChargeVersionsService` returns we'll need to identify if they are linked to a previous `billing_batch`. From there we'll need to locate the transactions and use them to generate news ones, that reverse what was previously done.

This PR creates the `ReverseBillingBatchLicencesService` which accepts a billing batch and set of licences, and returns the reversing transactions.