Create fetch billing invoice licence service
https://eaflood.atlassian.net/browse/WATER-3936

When we generate a supplementary bill run we are required to reverse the previous one. Essentially, when a licence is flagged for supplementary billing, for whatever reason, we reverse any previous bill runs and generate a new one.

With the exception of Annual billing, because we know any previous supplementary bill run will have reversed the one that came before, any new runs also just need to focus on the ones previous to them.

So, for each charge version `FetchChargeVersionsService` returns we'll need to identify if they are linked to a previous `billing_batch`. From there we'll need to locate the transactions and use them to generate news ones, that reverse what was previously done.

This PR creates the `FetchBillingInvoiceLicenceService` which accepts a `licenceId` and `financialYearEnding`. If it has been previously billed it returns the `billingInvoiceLicenceId` from the last time the licence was billed. The `billingInvoiceLicenceId` can then be used to retrieve the last set of transactions for the licence from the `billing_transactions` table.