Create Reverse Billing Batch Licences service (#167)
https://eaflood.atlassian.net/browse/WATER-3936

When we generate a supplementary bill run we are required to reverse the debits from the previous one. Essentially, when a licence is flagged for supplementary billing, for whatever reason, we reverse any previous bill runs and generate a new one.

With the exception of Annual billing, because we know any previous supplementary bill run will have reversed the debits on the one that came before, any new runs also just need to focus on the ones previous to them.

This is complicated, however, by the fact we then need to ignore transactions that cancel each other out. For example, Customer **ABC1** was billed 120 days for licence **LIC001** on the previous bill run. Our supplementary process calculates 120 days for the new charge version. These cancel each other out so should not be included.

So, for each charge version `FetchChargeVersionsService` returns we'll need to identify if they are linked to a previous `billing_batch`. From there we'll need to locate the transactions, determine if they cancel out what we've calculated and if not include them as credits in the supplementary bill run we're creating.

> Note - This PR originated to create a service to solve a part of this problem. It then became our 'SPIKE' branch and has culminated in all the changes needed. We _will_ be doing further cleanup and documentation after this change has been merged.